Un-nest AliasAnalysis Info and Graph to allow easier work with them from Java
