+++ engine/runtime-compiler/src/main/scala/org/enso/compiler/context/LocalScope.scala
@@ -1,12 +1,6 @@
 package org.enso.compiler.context
 
-import org.enso.compiler.pass.analyse.AliasAnalysis.Graph
-import org.enso.compiler.pass.analyse.AliasAnalysis.Graph.{
-  Id,
-  Occurrence,
-  Scope => AliasScope
-}
-import org.enso.compiler.pass.analyse.{AliasAnalysis, DataflowAnalysis}
+import org.enso.compiler.pass.analyse.{AliasAnalysisGraph, DataflowAnalysis}
 
 import scala.jdk.CollectionConverters._
 
@@ -33,19 +27,19 @@ import scala.jdk.CollectionConverters._
   */
 class LocalScope(
   final val parentScope: Option[LocalScope],
-  final val aliasingGraph: AliasAnalysis.Graph,
-  final val scope: AliasAnalysis.Graph.Scope,
+  final val aliasingGraph: AliasAnalysisGraph,
+  final val scope: AliasAnalysisGraph.Scope,
   final val dataflowInfo: DataflowAnalysis.Metadata,
-  final val flattenToParent: Boolean                  = false,
-  private val parentFrameSlotIdxs: Map[Graph.Id, Int] = Map()
+  final val flattenToParent: Boolean                               = false,
+  private val parentFrameSlotIdxs: Map[AliasAnalysisGraph.Id, Int] = Map()
 ) {
-  private lazy val localFrameSlotIdxs: Map[Graph.Id, Int] =
+  private lazy val localFrameSlotIdxs: Map[AliasAnalysisGraph.Id, Int] =
     gatherLocalFrameSlotIdxs()
 
   /** All frame slot indexes, including local and all the parents.
     * Useful for quick searching for [[FramePointer]] of parent scopes.
     */
-  private lazy val allFrameSlotIdxs: Map[Graph.Id, Int] =
+  private lazy val allFrameSlotIdxs: Map[AliasAnalysisGraph.Id, Int] =
     parentFrameSlotIdxs ++ localFrameSlotIdxs
 
   /** Creates a new child with a new aliasing scope.
@@ -62,7 +56,7 @@ class LocalScope(
     * @return a child of this scope
     */
   def createChild(
-    childScope: AliasScope,
+    childScope: AliasAnalysisGraph.Scope,
     flattenToParent: Boolean = false
   ): LocalScope = {
     new LocalScope(
@@ -83,7 +77,7 @@ class LocalScope(
     *           analysis.
     * @return the frame slot index for `id`.
     */
-  def getVarSlotIdx(id: Graph.Id): Int = {
+  def getVarSlotIdx(id: AliasAnalysisGraph.Id): Int = {
     assert(
       localFrameSlotIdxs.contains(id),
       "Cannot find " + id + " in " + localFrameSlotIdxs
@@ -98,7 +92,7 @@ class LocalScope(
     *           analysis
     * @return the frame pointer for `id`, if it exists
     */
-  def getFramePointer(id: Graph.Id): Option[FramePointer] = {
+  def getFramePointer(id: AliasAnalysisGraph.Id): Option[FramePointer] = {
     aliasingGraph
       .defLinkFor(id)
       .flatMap { link =>
@@ -126,7 +120,7 @@ class LocalScope(
     *         indexes in the frame. Takes into account all the
     *         internal slots, that are prepended to every frame.
     */
-  private def gatherLocalFrameSlotIdxs(): Map[Id, Int] = {
+  private def gatherLocalFrameSlotIdxs(): Map[AliasAnalysisGraph.Id, Int] = {
     scope.allDefinitions.zipWithIndex.map { case (definition, i) =>
       definition.id -> (i + LocalScope.internalSlotsSize)
     }.toMap
@@ -139,13 +133,13 @@ class LocalScope(
     */
   private def flattenBindingsWithLevel(
     level: Int
-  ): Map[Graph.Symbol, FramePointer] = {
-    var parentResult: Map[Graph.Symbol, FramePointer] = parentScope
+  ): Map[AliasAnalysisGraph.Symbol, FramePointer] = {
+    var parentResult: Map[AliasAnalysisGraph.Symbol, FramePointer] = parentScope
       .flatMap(scope => Some(scope.flattenBindingsWithLevel(level + 1)))
       .getOrElse(Map())
 
     scope.occurrences.foreach {
-      case x: Occurrence.Def =>
+      case x: AliasAnalysisGraph.Occurrence.Def =>
         parentResult += x.symbol -> new FramePointer(
           level,
           allFrameSlotIdxs(x.id)
@@ -166,7 +160,7 @@ object LocalScope {
     * @return a defaulted local scope
     */
   def root: LocalScope = {
-    val graph = new AliasAnalysis.Graph
+    val graph = new AliasAnalysisGraph
     new LocalScope(
       None,
       graph,
