+++ engine/runtime/src/main/scala/org/enso/interpreter/runtime/IrToTruffle.scala
@@ -9,71 +9,24 @@ import org.enso.compiler.core.CompilerError
 import org.enso.compiler.core.ConstantsNames
 import org.enso.compiler.core.Implicits.AsMetadata
 import org.enso.compiler.core.IR
-import org.enso.compiler.core.ir.{
-  CallArgument,
-  DefinitionArgument,
-  Empty,
-  Expression,
-  Function,
-  IdentifiedLocation,
-  Literal,
-  Module,
-  Name,
-  Pattern,
-  `type`,
-  Type => Tpe
-}
+import org.enso.compiler.core.ir.{CallArgument, DefinitionArgument, Empty, Expression, Function, IdentifiedLocation, Literal, Module, Name, Pattern, `type`, Type => Tpe}
 import org.enso.compiler.core.ir.module.scope.Definition
 import org.enso.compiler.core.ir.module.scope.definition
 import org.enso.compiler.core.ir.module.scope.Import
 import org.enso.compiler.core.ir.module.scope.imports
 import org.enso.compiler.core.ir.Name.Special
-import org.enso.compiler.core.ir.expression.{
-  errors,
-  Application,
-  Case,
-  Comment,
-  Error,
-  Foreign,
-  Operator,
-  Section
-}
-import org.enso.compiler.data.BindingsMap.{
-  ExportedModule,
-  ResolvedConstructor,
-  ResolvedModule
-}
+import org.enso.compiler.core.ir.expression.{Application, Case, Comment, Error, Foreign, Operator, Section, errors}
+import org.enso.compiler.data.BindingsMap.{ExportedModule, ResolvedConstructor, ResolvedModule}
 import org.enso.compiler.data.{BindingsMap, CompilerConfig}
 import org.enso.compiler.exception.BadPatternMatch
-import org.enso.compiler.pass.analyse.AliasAnalysis.Graph.{Scope => AliasScope}
-import org.enso.compiler.pass.analyse.AliasAnalysis.{Graph => AliasGraph}
-import org.enso.compiler.pass.analyse.{
-  AliasAnalysis,
-  BindingAnalysis,
-  DataflowAnalysis,
-  TailCall
-}
-import org.enso.compiler.pass.resolve.{
-  ExpressionAnnotations,
-  GenericAnnotations,
-  GlobalNames,
-  MethodDefinitions,
-  Patterns,
-  TypeNames,
-  TypeSignatures
-}
+import org.enso.compiler.pass.analyse.AliasAnalysisGraph.{Scope => AliasScope}
+import org.enso.compiler.pass.analyse.{AliasAnalysis, AliasAnalysisGraph, AliasAnalysisInfo, BindingAnalysis, DataflowAnalysis, TailCall}
+import org.enso.compiler.pass.resolve.{ExpressionAnnotations, GenericAnnotations, GlobalNames, MethodDefinitions, Patterns, TypeNames, TypeSignatures}
 import org.enso.interpreter.node.callable.argument.ReadArgumentNode
 import org.enso.interpreter.node.callable.argument.ReadArgumentCheckNode
-import org.enso.interpreter.node.callable.function.{
-  BlockNode,
-  CreateFunctionNode
-}
+import org.enso.interpreter.node.callable.function.{BlockNode, CreateFunctionNode}
 import org.enso.interpreter.node.callable.thunk.{CreateThunkNode, ForceNode}
-import org.enso.interpreter.node.callable.{
-  ApplicationNode,
-  InvokeCallableNode,
-  SequenceLiteralNode
-}
+import org.enso.interpreter.node.callable.{ApplicationNode, InvokeCallableNode, SequenceLiteralNode}
 import org.enso.interpreter.node.controlflow.caseexpr._
 import org.enso.interpreter.node.expression.builtin.interop.syntax.HostValueToEnsoNode
 import org.enso.interpreter.node.expression.builtin.BuiltinRootNode
@@ -81,30 +34,16 @@ import org.enso.interpreter.node.expression.constant._
 import org.enso.interpreter.node.expression.foreign.ForeignMethodCallNode
 import org.enso.interpreter.node.expression.literal.LiteralNode
 import org.enso.interpreter.node.scope.{AssignmentNode, ReadLocalVariableNode}
-import org.enso.interpreter.node.{
-  BaseNode,
-  ClosureRootNode,
-  ConstantNode,
-  MethodRootNode,
-  ExpressionNode => RuntimeExpression
-}
+import org.enso.interpreter.node.{BaseNode, ClosureRootNode, ConstantNode, MethodRootNode, ExpressionNode => RuntimeExpression}
 import org.enso.interpreter.runtime.EnsoContext
 import org.enso.interpreter.runtime.callable
-import org.enso.interpreter.runtime.callable.argument.{ArgumentDefinition}
+import org.enso.interpreter.runtime.callable.argument.ArgumentDefinition
 import org.enso.interpreter.runtime.data.atom.{Atom, AtomConstructor}
-import org.enso.interpreter.runtime.callable.function.{
-  FunctionSchema,
-  Function => RuntimeFunction
-}
-import org.enso.interpreter.runtime.callable.{
-  UnresolvedConstructor,
-  UnresolvedConversion,
-  UnresolvedSymbol,
-  Annotation => RuntimeAnnotation
-}
+import org.enso.interpreter.runtime.callable.function.{FunctionSchema, Function => RuntimeFunction}
+import org.enso.interpreter.runtime.callable.{UnresolvedConstructor, UnresolvedConversion, UnresolvedSymbol, Annotation => RuntimeAnnotation}
 import org.enso.interpreter.runtime.data.Type
 import org.enso.interpreter.runtime.data.text.Text
-import org.enso.interpreter.runtime.scope.{ModuleScope}
+import org.enso.interpreter.runtime.scope.ModuleScope
 import org.enso.interpreter.{Constants, EnsoLanguage}
 
 import java.math.BigInteger
@@ -259,7 +198,7 @@ class IrToTruffle(
               AliasAnalysis,
               "No root scope on an atom definition."
             )
-            .unsafeAs[AliasAnalysis.Info.Scope.Root]
+            .unsafeAs[AliasAnalysisInfo.Scope.Root]
 
           val dataflowInfo = atomDefn.unsafeGetMetadata(
             DataflowAnalysis,
@@ -290,7 +229,7 @@ class IrToTruffle(
                 AliasAnalysis,
                 "No occurrence on an argument definition."
               )
-              .unsafeAs[AliasAnalysis.Info.Occurrence]
+              .unsafeAs[AliasAnalysisInfo.Occurrence]
             val slotIdx = localScope.getVarSlotIdx(occInfo.id)
             argDefs(idx) = arg
             val readArg =
@@ -359,7 +298,7 @@ class IrToTruffle(
           s"Missing scope information for method " +
           s"`${methodDef.typeName.map(_.name + ".").getOrElse("")}${methodDef.methodName.name}`."
         )
-        .unsafeAs[AliasAnalysis.Info.Scope.Root]
+        .unsafeAs[AliasAnalysisInfo.Scope.Root]
       val dataflowInfo = methodDef.unsafeGetMetadata(
         DataflowAnalysis,
         "Method definition missing dataflow information."
@@ -595,7 +534,7 @@ class IrToTruffle(
                               scopeElements.init
                                 .mkString(Constants.SCOPE_SEPARATOR)
                             )
-                            .unsafeAs[AliasAnalysis.Info.Scope.Root]
+                            .unsafeAs[AliasAnalysisInfo.Scope.Root]
                           val dataflowInfo = annotation.unsafeGetMetadata(
                             DataflowAnalysis,
                             "Missing dataflow information for annotation " +
@@ -672,7 +611,7 @@ class IrToTruffle(
           s"Missing scope information for conversion " +
           s"`${methodDef.typeName.map(_.name + ".").getOrElse("")}${methodDef.methodName.name}`."
         )
-        .unsafeAs[AliasAnalysis.Info.Scope.Root]
+        .unsafeAs[AliasAnalysisInfo.Scope.Root]
       val dataflowInfo = methodDef.unsafeGetMetadata(
         DataflowAnalysis,
         "Method definition missing dataflow information."
@@ -1012,7 +951,7 @@ class IrToTruffle(
       */
     def this(
       scopeName: String,
-      graph: AliasGraph,
+      graph: AliasAnalysisGraph,
       scope: AliasScope,
       dataflowInfo: DataflowAnalysis.Metadata
     ) = {
@@ -1122,7 +1061,7 @@ class IrToTruffle(
             AliasAnalysis,
             "Missing scope information on block."
           )
-          .unsafeAs[AliasAnalysis.Info.Scope.Child]
+          .unsafeAs[AliasAnalysisInfo.Scope.Child]
 
         val childFactory = this.createChild("suspended-block", scopeInfo.scope)
         val childScope   = childFactory.scope
@@ -1232,7 +1171,7 @@ class IrToTruffle(
           AliasAnalysis,
           "No scope information on a case branch."
         )
-        .unsafeAs[AliasAnalysis.Info.Scope.Child]
+        .unsafeAs[AliasAnalysisInfo.Scope.Child]
 
       val childProcessor = this.createChild("case_branch", scopeInfo.scope)
 
@@ -1601,7 +1540,7 @@ class IrToTruffle(
           AliasAnalysis,
           "Binding with missing occurrence information."
         )
-        .unsafeAs[AliasAnalysis.Info.Occurrence]
+        .unsafeAs[AliasAnalysisInfo.Occurrence]
 
       currentVarName = binding.name.name
 
@@ -1625,7 +1564,7 @@ class IrToTruffle(
     ): RuntimeExpression = {
       val scopeInfo = function
         .unsafeGetMetadata(AliasAnalysis, "No scope info on a function.")
-        .unsafeAs[AliasAnalysis.Info.Scope.Child]
+        .unsafeAs[AliasAnalysisInfo.Scope.Child]
 
       if (function.body.isInstanceOf[Function]) {
         throw new CompilerError(
@@ -1665,7 +1604,7 @@ class IrToTruffle(
               AliasAnalysis,
               "No occurrence on variable usage."
             )
-            .unsafeAs[AliasAnalysis.Info.Occurrence]
+            .unsafeAs[AliasAnalysisInfo.Occurrence]
 
           val framePointer = scope.getFramePointer(useInfo.id)
           val global       = name.getMetadata(GlobalNames)
@@ -1946,7 +1885,7 @@ class IrToTruffle(
                 AliasAnalysis,
                 "No occurrence on an argument definition."
               )
-              .unsafeAs[AliasAnalysis.Info.Occurrence]
+              .unsafeAs[AliasAnalysisInfo.Occurrence]
 
             val slotIdx = scope.getVarSlotIdx(occInfo.id)
             val readArg =
@@ -2176,7 +2115,7 @@ class IrToTruffle(
               AliasAnalysis,
               "No scope attached to a call argument."
             )
-            .unsafeAs[AliasAnalysis.Info.Scope.Child]
+            .unsafeAs[AliasAnalysisInfo.Scope.Child]
 
           val shouldCreateClosureRootNode = value match {
             case _: Name           => false
